requirements.txt
google-generativeai
google-auth
google-auth-httplib2
google-api-python-client
PyPDF2
requests
pandas